# VRTWarpedDataset: add an optimized IRasterIO() implementation