This specialized implementation of IRasterIO() will be faster than using the VRTWarpedRasterBand::IReadBlock() method in situations where
- a large enough chunk of data is requested at once
- and multi-threaded warping is enabled (it only kicks in if the warped chunk is large enough) and/or when reading the source dataset is multi-threaded (e.g JP2KAK or JP2OpenJPEG driver).

e.g. with a 3909x3154 4-band UInt16 input JPEG2000 source dataset:

```
$ gdalwarp -overwrite -t_srs EPSG:4326 -r cubic -wo NUM_THREADS=12 input.jp2 test.vrt
Creating output file that is 3909P x 3154L.
```

Now:
```
$ time GDAL_SKIP=JP2KAK,JP2ECW gdal_translate test.vrt out.tif -co tiled=yes
Input file size is 3909, 3154
0...10...20...30...40...50...60...70...80...90...100 - done.

real    0m1,624s
user    0m11,910s
sys     0m0,270s
```

Before (or disabling the optimization):
```
$ time GDAL_VRT_WARP_USE_DATASET_RASTERIO=NO GDAL_SKIP=JP2KAK,JP2ECW gdal_translate test.vrt out.tif -co tiled=yes
Input file size is 3909, 3154
0...10...20...30...40...50...60...70...80...90...100 - done.

real    0m4,115s
user    0m9,934s
sys     0m0,218s
```